requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
websockets
python-multipart
